requirements.txt
Django==3.0
Pillow==7.0.0
django-jsonfield==1.4.0